moved Graph from baseplotter to vizualization
